# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/users": {
      "post": {
        "summary": "Register User",
        "operationId": "register_user_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/cities": {
      "post": {
        "summary": "Add City",
        "operationId": "add_city_users__user_id__cities_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CityCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Cities",
        "operationId": "get_cities_users__user_id__cities_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Cities Users  User Id  Cities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/cities/{city_name}/forecast": {
      "get": {
        "summary": "Get Forecast",
        "operationId": "get_forecast_users__user_id__cities__city_name__forecast_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "city_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "City Name"
            }
          },
          {
            "name": "time",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Format HH:MM",
              "title": "Time"
            },
            "description": "Format HH:MM"
          },
          {
            "name": "params",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Params"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/forecast": {
      "get": {
        "summary": "Get Current Weather",
        "operationId": "get_current_weather_forecast_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Lon"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentWeatherResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CityCreate": {
        "properties": {
          "city_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "City Name"
          },
          "latitude": {
            "type": "number",
            "maximum": 90,
            "minimum": -90,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180,
            "minimum": -180,
            "title": "Longitude"
          }
        },
        "type": "object",
        "required": [
          "city_name",
          "latitude",
          "longitude"
        ],
        "title": "CityCreate"
      },
      "CityResponse": {
        "properties": {
          "city_id": {
            "type": "integer",
            "title": "City Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "OK"
          }
        },
        "type": "object",
        "required": [
          "city_id"
        ],
        "title": "CityResponse"
      },
      "CurrentWeatherResponse": {
        "properties": {
          "temperature": {
            "type": "number",
            "title": "Temperature"
          },
          "wind_speed": {
            "type": "number",
            "title": "Wind Speed"
          },
          "pressure": {
            "type": "number",
            "title": "Pressure"
          }
        },
        "type": "object",
        "required": [
          "temperature",
          "wind_speed",
          "pressure"
        ],
        "title": "CurrentWeatherResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "UserCreate": {
        "properties": {
          "user_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 5,
            "title": "User Name"
          }
        },
        "type": "object",
        "required": [
          "user_name"
        ],
        "title": "UserCreate"
      },
      "UserResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}